requirements.txt
httpx==0.27.0
beautifulsoup4==4.12.3
python-dateutil==2.9.0.post0
icalendar==5.0.12
pytz==2024.1
PyYAML==6.0.2
